Standardize the case of the `sessionId` and `sessionStart` keys
Currently, the `sessionId` and `sessionStart` keys, which are used in the events
context to pass session information, are incorrectly referenced in some cases by
using the first character in upper case.

This commit fixes the issue by ensuring that these keys are always referenced
with the first character in lower case.