Fix HTTP and Strategy tests
This commit fixes the tests that were failing when executed through the `CI`
Github workflow introduced with commit a56d5d6.

Specifically, the commit fixes the Strategy tests by adding cleanup operations
before and after the execution of all the class tests. In addition, it disables
the tests that were testing HTTP calls to the endpoints until we have our own
domain to use specifically for the tests.